requirements.txt
# DataRobot API Code Generation Agent Dependencies

# Core DataRobot SDK
datarobot>=3.5.0

# LLM Provider APIs
anthropic>=0.34.0
openai>=1.54.0
google-generativeai>=0.8.0

# Vector Search & Embeddings
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0

# Text Processing & Search
scikit-learn>=1.5.0
rank-bm25>=0.2.2
nltk>=3.8.1

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# HTTP & Networking
requests>=2.31.0
aiohttp>=3.9.0

# File Processing
PyYAML>=6.0.1
beautifulsoup4>=4.12.0
lxml>=5.2.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0